+++ .eleventy.js
@@ -73,7 +73,7 @@ const {getDefaultUrl, getRelativePath} = require('./src/site/_filters/urls');
 const {memoize, findByUrl} = require('./src/site/_filters/find-by-url');
 const pathSlug = require('./src/site/_filters/path-slug');
 const algoliaIndexable = require('./src/site/_filters/algolia-indexable');
-const algoliaItem = require('./src/site/_filters/algolia-item');
+const {algoliaItem} = require('./src/site/_filters/algolia-item');
 const containsTag = require('./src/site/_filters/contains-tag');
 const expandAuthors = require('./src/site/_filters/expand-authors');
 const githubLink = require('./src/site/_filters/github-link');
