+++ src/site/_filters/algolia-item.js
@@ -20,7 +20,17 @@ const {getDefaultUrl} = require('./urls');
 const {defaultLocale} = require('../_data/site');
 const {generateImgixSrc} = require('../_includes/components/Img');
 const {findByUrl} = require('./find-by-url');
-const {getPostParentUrl} = require('./get-post-parent-url');
+
+/**
+ * Returns the parent URL of a post. For example:
+ * /en/blog/example-post -> /en/blog/
+ *
+ * @param {EleventyCollectionItem} post
+ * @returns {string}
+ */
+function getPostParentUrl(post) {
+  return `${post.data.page.filePathStem.split('/').slice(0, -2).join('/')}/`;
+}
 
 /**
  * @param {EleventyCollectionItem} post
@@ -28,6 +38,20 @@ const {getPostParentUrl} = require('./get-post-parent-url');
  */
 function algoliaItem(post) {
   const data = post.data;
+
+  let parent;
+  try {
+    parent = findByUrl(getPostParentUrl(post));
+  } catch (e) {
+    if (
+      e instanceof Error &&
+      e.message === 'No collection has been memoized yet.'
+    ) {
+      parent = null;
+    } else {
+      throw e;
+    }
+  }
 
   return {
     content: striptags(/** @type {string} */ (post.templateContent)),
@@ -44,8 +68,8 @@ function algoliaItem(post) {
     title: data.renderData?.title || data.title,
     updatedOn: data.updated,
     url: getDefaultUrl(post.url),
-    parentTitle: findByUrl(getPostParentUrl(post))?.data?.title,
+    parentTitle: parent?.data?.title,
   };
 }
 
-module.exports = algoliaItem;
+module.exports = {algoliaItem, getPostParentUrl};
